Fix archive template
Because of how scoping works in Jinja, the year would be printed for
every entry instead of once, as was the intention. Now it looks at the
previous article's year and directly compares that.

This causes the first article to always be compared to the last but
that's OK because for the first article we'll always want to print the
year.