Add support for dark mode
Adds a color scheme and general support for dark mode. This requires
inlining the svg image contents in the html so they can be given CSS
colors. To not change too much about the general setup of the website,
sa Python script can generate the svg content in an HTML page.

For simplicity and because the icon set doesn't change that frequently,
the script doesn't run every time but can still be manually invoked when
needed.