Add AI disclaimer to archive list
This defines the archive page in the content folder only for the
disclaimer. This is better than before where the entire archive page
lived in the template, but it is a little weird that part of the
page lives in the template and part in the markdown. Ideally pelican
supports something like `{BLOG_ARCHIVE_LIST}` but that doesn't exist
today hence the hybrid.

An alternative would have been to put this sentence in the template but
that didn't feel as clean.